Add traces routed by hand
Previously the diodes were connected from the columns to the switches.
Both the ErgoDash and the Corne have them behind the switches connected
to the rows. Since I positioned the diodes on the PCB like they are in
the ErgoDash and Corne, the routing was much more difficult than intended.

This moves the diodes behind the switches and also moves them around a
bit on the PCB. This allowed me to hand route all traces easily, so this
time no autorouter has been involved.

The via count is rather large (156). I opted to rather have ONE trace
and then interconnect each component with vias itself. An alternative
could be to run traces for rows and columns on both layers, but then
there still would be a lot of conflicting paths requiring vias anyway.

I also removed both GND zones. There is nothing on the board that needs
ground except the nice!nano and the battery and I didn't want to start
adding complicated routing just to get GND zones on the whole board. I
left one GND zone in the nice!nano area since it seems reasonable there.